Move and add tests for nested validation
